split exception
